+++ src/timecraft.js
@@ -1,85 +1,82 @@
-var kernel_paths = ['/kernels/lsk/naif0012.tls','/kernels/spk/de425s.bsp','/kernels/pck/pck00010.tpc',];
-var timecraftjs;
-//Load the required javascript files
-timecraftjs = {"is_ready":false};
+import Module from './cspice.js';
+import * as Spice from './spice.js';
 
-//set up spice different ways depending on if in a browser or not
-if(typeof window == "undefined"){
-	//If in node
-    timecraft = require("timecraftjs/spice.js");
-	var cspice = require("timecraftjs/cspice.js");
-	var fs = require("fs");
-    timecraft.SPICE.setup(cspice,fs);
-	for(var i = 0;i < kernel_paths.length;i++){
-        timecraft.SPICE.node_furnish(kernel_paths[i]);
-	}
-    timecraft = timecraft.SPICE;
-    timecraft.is_ready = true;
-} else {
-	//If in a browse
+let bufferFileCount = 0;
 
+const FS = Module.get_fs();
 
+function mkdirRecursive(parts) {
+    const tempParts = parts.slice();
 
-	//sp = {"SPICE":SPICE};
-	//SPICE.furnish("./kernels/lsk/naif0012.tls")
-	var Module = {
-
-
-		// The functions to run after loading
-		postRun: [function(){
-			SPICE.setup(Module,FS);
-			/*
-				Read files into the program from the virtual file system
-				SPICE.furnish(['kernels/pck/pck00010.tpc','kernels/lsk/naif0010.tls','kernels/spk/planets/de432s.bsp']) would also have worked.
-			*/
-			if(typeof preload_file_data == "undefined"){
-				SPICE.furnish_via_xhr_request(kernel_paths,
-				function(){
-				},
-				function(){
-                    timecraft = SPICE;
-                    timecraft.is_ready = true;
-					var evnt = new CustomEvent("timecraftready");
-					window.dispatchEvent(evnt);
-				});
-			} else {
-				for(var i = 0;i < preload_file_data.length;i++){
-					var full_path = preload_file_data[i].path;
-					SPICE.furnish_via_preload_file_data(full_path,preload_file_data[i].buffer)
-				}
-                timecraft = SPICE;
-                timecraft.is_ready = true;
-				preload_file_data = null;
-				var evnt = new CustomEvent("timecraftready");
-				window.dispatchEvent(evnt);
-			}
-		}
-		],
-		// Capture stdout
-		print: function() {
-			text = Array.prototype.slice.call(arguments).join(' ');
-			if(timecraft.is_ready){
-				if(timecraft.report_then_reset && timecraft.failed() == 1){
-					var errstr = timecraft.getmsg("LONG");
-					//console.log(text);
-                    timecraft.reset();
-					console.error(errstr);
-				} else {
-					console.log(text);
-				}
-			}
-		},
-		// Capture stderr
-		printErr: function() {
-			text = Array.prototype.slice.call(arguments).join(' ');
-			if(text.slice(0,8) === 'pre-main') return;
-			console.error(text);
-		},
-
-
-
-
-	};
+    let currPath = '';
+    while (tempParts.length) {
+        if (currPath !== '') {
+            currPath += '/';
+        }
+        currPath += tempParts.shift();
+        try {
+            FS.mkdir(currPath);
+        } catch (e) {
+            // do nothing
+        }
+    }
 }
 
-if(typeof window == "undefined") module.exports = timecraft;
+// file system manipulation
+// https://emscripten.org/docs/api_reference/Filesystem-API.html
+export function prepareFileFromBuffer(path, buffer) {
+    if (buffer instanceof ArrayBuffer) {
+        buffer = new Uint8Array(buffer);
+    }
+
+    const parts = path.split(/[\\/]/g);
+    parts.pop();
+    mkdirRecursive(parts);
+
+    FS.writeFile(path, buffer, { encoding: 'binary' });
+}
+
+export function removeFile(path) {
+    FS.unlink(path);
+}
+
+// loading kernels
+export function loadKernel(path) {
+    Spice.furnsh(path);
+}
+
+export function loadKernelFromBuffer(buffer) {
+    const path = `__buffer_files__/buffer_${ bufferFileCount }.bin`;
+    bufferFileCount++;
+
+    prepareFileFromBuffer(path, buffer);
+    loadKernel(path);
+}
+
+// unloading kernel
+export function unloadKernel(key) {
+    Spice.unload(key);
+}
+
+// Chronos CLI
+// https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/ug/chronos.html
+
+export function chronos(inptim, cmdlin) {
+
+    const outtim_ptr = Module._malloc(256);
+    const intptr = Module._malloc(8);
+
+    Module.setValue(intptr, 1, 'i32');
+    Module.ccall(
+        'cronos_',
+        'number',
+        ['string', 'number', 'string', 'number', 'number', 'number', 'number'],
+        [cmdlin, intptr, inptim, outtim_ptr, cmdlin.length, inptim.length, 256],
+    );
+
+    const ret = Module.Pointer_stringify(outtim_ptr);
+    Module._free(outtim_ptr);
+
+    return ret;
+
+}
